+++ ISS 3D Simulation/ISS/ISS.prefab.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: f0acf4e1f22ac4961a07e4d875a43ec6
+timeCreated: 1504844546
+licenseType: Free
+NativeFormatImporter:
+  mainObjectFileID: 100100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
