+++ ISS 3D Simulation/ISS/PositionISS.cs
@@ -2,81 +2,40 @@
 using System.Collections.Generic;
 using UnityEngine;
 
-using Boomlagoon.JSON;
-
 public class PositionISS : MonoBehaviour {
 
   public GameObject earthParent;
   public GameObject earth;
 
-  private string issApiUrl = "http://api.open-notify.org/iss-now.json";
-  private int positionsToRecord = 3;
-  private Vector2[] recentISSPositions = new Vector2[5];
-  private IEnumerator coroutine;
+  private GetISSData getDataScript;
+  public int positionsToRecord = 3;
 
   void Start ()
   {
-    InvokeRepeating("CallISSAPI", 0f, 1.0f);
-    InitialISSPosition();
+    getDataScript = GetComponent<GetISSData>();
+    earthParent.transform.localScale = new Vector3(.25f, .25f, .25f);
+    transform.localScale = new Vector3(.001f, .001f, .001f);
     InvokeRepeating("SetISSPosition", positionsToRecord, 1.0f);
   }
 
-  void CallISSAPI ()
+  public void AnimateToInitialISSPosition (Vector2 position)
   {
-    if(coroutine != null) StopCoroutine(coroutine);
-    coroutine = GetISSDataFromAPI();
-    StartCoroutine(coroutine);
-  }
-
-  IEnumerator GetISSDataFromAPI ()
-  {
-    WWW www = new WWW (issApiUrl);
-    yield return www;
-    if (www.text != "") {
-      Vector2 position = ParseISSJSONData(www.text);
-      AddPositionToArray(position);
-    } else {
-      Debug.Log("Error: " + www.error);
-    }
-  }
-
-  Vector2 ParseISSJSONData (string apiText)
-  {
-    JSONObject json = JSONObject.Parse(apiText);
-    JSONObject iss_position = json.GetObject("iss_position");
-    float longitude = float.Parse(iss_position.GetString("longitude"));
-    float latitude = float.Parse(iss_position.GetString("latitude"));
-    return new Vector2(longitude, latitude);
-  }
-
-  void InitialISSPosition () {
-
+    iTween.ScaleTo(earthParent, iTween.Hash("scale", Vector3.one,"time", 4.0f,"delay", 2.0f,"easeType", "easeInOutSine"));
+    iTween.ScaleTo(gameObject, iTween.Hash("scale", Vector3.one, "time", 1.0f, "delay", 5.0f, "easeType", "easeOutSine"));
+    // Subtract the time it took to load the initial data. Otherwise, you get overlapping animations.
+    AnimateISS(position[0], position[1], positionsToRecord - Time.time, "easeInOutSine");
   }
 
   void SetISSPosition ()
   {
-    float latitude = recentISSPositions[positionsToRecord-1][1];
-    float longitude = recentISSPositions[positionsToRecord-1][0];
-
-    iTween.RotateTo(earthParent, iTween.Hash(
-      "rotation", new Vector3(0,0,latitude),
-      "time", 1.0f,
-      "easeType", "linear"
-    ));
-
-    iTween.RotateTo(earth, iTween.Hash(
-      "rotation", new Vector3(0, longitude-90, 0),
-      "time", 1.0f,
-      "easeType", "linear",
-      "islocal", true
-    ));
+    float latitude = getDataScript.recentISSPositions[positionsToRecord-1][1];
+    float longitude = getDataScript.recentISSPositions[positionsToRecord-1][0];
+    AnimateISS(longitude, latitude, 1.0f, "linear");
   }
 
-  void AddPositionToArray (Vector2 newItem)
+  void AnimateISS (float longitude, float latitude, float duration, string easeType)
   {
-    for (int i = positionsToRecord-1; i > 0; i--) {
-      recentISSPositions[i] = recentISSPositions[i-1];
-    }
-    recentISSPositions[0] = newItem;
+    iTween.RotateTo(earthParent, iTween.Hash("rotation", new Vector3(0,0,latitude), "time", duration, "easeType", easeType));
+    iTween.RotateTo(earth, iTween.Hash("rotation", new Vector3(0, longitude-90, 0), "time", duration, "easeType", easeType, "islocal", true));
   }
 }
